Publish ES modules build
The package.json contains a `module` field, but the `es/index.js` file
is not published in the package. While bundlers like Webpack will just
fall back to `main`, eslint-plugin-import complains that the path does
not resolve to a real file.

This patch runs the existing `build:es` npm script before publishing and
adds the `es` folder to the package, so that you can `import`
react-google-button in ES modules bundlers.

I moved the babel config into a custom preset (sort of), because I don't
think you can override options for a single preset using the JSON env
syntax. When BABEL_ENV=es, ES modules syntax is not compiled.